+++ src/lib/AAHelpers.mjs
@@ -5,6 +5,7 @@ export class AAHelpers {
 
   static evaluateCustomCheck(token, check) {
     try {
+      // these are exposed here so they can by used in the custom check/eval
       // eslint-disable-next-line no-unused-vars
       const actor = token.actor;
       // eslint-disable-next-line no-unused-vars
@@ -79,17 +80,36 @@ export class AAHelpers {
         return AAHelpers.typeCheck4e(canvasToken, type);
     }
   }
+
+  static CheckTypes(canvasToken, types) {
+    if (!types || types === "") return true;
+    const typesArray = Array.isArray(types)
+      ? types
+      : types?.toLowerCase().replaceAll(",", ";").split(";").map((t) => t.trim()) ?? [];
+
+    let match = false;
+    for (const type of typesArray) {
+      if (AAHelpers.CheckType(canvasToken, type)) {
+        match = true;
+        break;
+      }
+    }
+    return match;
+  }
+
   static typeCheck5e(canvasToken, type) {
+    if (type?.trim() === "any") return true;
     const systemData = canvasToken?.actor?.system;
-    let tokenType;
+    let tokenTypes;
     switch (canvasToken.actor.type) {
       case "npc":
         {
           try {
-            tokenType = [
+            tokenTypes = Array.from(new Set([
               systemData?.details.type.value,
-              systemData?.details.type.custom
-            ];
+              systemData?.details.type.subtype,
+              systemData?.details.type.custom,
+            ])).filter((t) => t);
           } catch (error) {
             Logger.error("ActiveAuras: the token has an unreadable type", canvasToken);
           }
@@ -99,12 +119,15 @@ export class AAHelpers {
         {
           try {
             if (game.system.id === "sw5e") {
-              tokenType = [systemData?.details.species.toLowerCase()];
+              tokenTypes = [systemData?.details.species.toLowerCase()];
             } else{
-              tokenType = [
+              tokenTypes = Array.from(new Set([
+                (systemData?.details?.race?.name ?? systemData?.details?.race)?.toLocaleLowerCase(),
                 (systemData?.details?.race?.name ?? systemData?.details?.race)?.toLocaleLowerCase().replace("-", " ").split(" "),
-                systemData?.details.type?.value?.toLocaleLowerCase().replace("-", " ").split(" "),
-              ].flat();
+                systemData?.details.type?.value?.toLocaleLowerCase(),
+                systemData?.details.type?.subtype?.toLocaleLowerCase(),
+                systemData?.details.type?.custom?.toLocaleLowerCase(),
+              ].flat())).filter((t) => t);
             }
           } catch (error) {
             Logger.error("ActiveAuras: the token has an unreadable type", canvasToken);
@@ -115,21 +138,24 @@ export class AAHelpers {
       case "vehicle":
         return;
     }
-    let humanoidRaces;
-    if (game.system.id === "sw5e") {
-      humanoidRaces = CONSTANTS.SW5E_HUMANOID_RACES;
-    } else humanoidRaces = CONSTANTS.HUMANOID_RACES;
 
-    if (tokenType.includes(type)) return true;
+    if (tokenTypes.includes(type)) return true;
 
-    for (let x of tokenType) {
+    // remaining humanoid checks only npcs in 5e or all in sw5e
+    if (type.trim() !== "humanoid") return false;
+    if (canvasToken.actor.type !== "character" && game.system.id === "dnd5e") return false;
+    const humanoidRaces = game.system.id === "sw5e"
+      ? CONSTANTS.SW5E_HUMANOID_RACES
+      : CONSTANTS.HUMANOID_RACES;
+
+    let match = false;
+    for (const x of tokenTypes) {
       if (humanoidRaces.includes(x)) {
-        tokenType = "humanoid";
-        continue;
+        match = true;
+        break;
       }
     }
-    if (tokenType === type || tokenType === "any") return true;
-    return false;
+    return match;
   }
 
   static typeCheckSWADE(canvasToken, type) {